Flyttar henting av person frå pdl med toggling og alt ut i eiga klasse, for å kunne gjenbruke denne enkelt
